Move code for both OCLC APIs into a new package.
