+++ bin/oclc_linked_data_coverage
@@ -6,7 +6,7 @@ bin_dir = os.path.split(__file__)[0]
 package_dir = os.path.join(bin_dir, "..")
 sys.path.append(os.path.abspath(package_dir))
 
-from oclc import LinkedDataCoverageProvider
+from oclc.linked_data import LinkedDataCoverageProvider
 from core.scripts import RunCoverageProviderScript
 
 RunCoverageProviderScript(LinkedDataCoverageProvider).run()
