+++ canonicalize.py
@@ -4,7 +4,7 @@ import os
 import re
 
 from nose.tools import set_trace
-from oclc import OCLCLinkedData
+from oclc.linked_data import OCLCLinkedData
 from viaf import VIAFClient, MockVIAFClient
 
 from core.model import (
