+++ scripts.py
@@ -44,8 +44,8 @@ from core.scripts import (
 from core.util.permanent_work_id import WorkIDCalculator
 from core.util.personal_names import contributor_name_match_ratio
 
-from oclc import LinkedDataCoverageProvider
-from oclc_classify import OCLCClassifyCoverageProvider
+from oclc.linked_data import LinkedDataCoverageProvider
+from oclc.classify import OCLCClassifyCoverageProvider
 from viaf import VIAFClient
